requirements.txt
google-genai
requests
urllib3
beautifulsoup4